+++ tests/Web.suite.yml
@@ -1,4 +1,4 @@
-actor: AcceptanceTester
+actor: WebTester
 extensions:
   enabled:
     - Codeception\Extension\RunProcess:
@@ -8,4 +8,3 @@ modules:
   enabled:
     - PhpBrowser:
         url: http://127.0.0.1:8080
-  step_decorators: ~
